backend/pyproject.toml
[project]
name = "study-buddy-backend"
version = "0.1.0"
description = "GenAI Backend for Personalized Study Buddy"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.29",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.6",
    "python-multipart>=0.0.9",
    "pymupdf>=1.24.1",
    "sentence-transformers>=2.6.1",
    "groq>=0.5.0",
    "httpx>=0.27.0",
    "google-auth>=2.29.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.2.1",
    "requests>=2.31.0",
    "email-validator>=2.1.0",
    "neo4j>=5.19.0",
    "fastapi-mail>=1.4.1",
    "ddgs>=9.10.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "black>=24.3.0",
    "isort>=5.13.2",
]
